+++ EggIncubator_HW41/Core/Src/string_num.c
@@ -14,10 +14,10 @@
 //	   *fractional_part=(int)((a-(double)*integer_part)*10.0);
 //}
 char ftoa_str[16];
-char * ftoa(float a )
+char * ftoa(double a )
 {
 	uint8_t integer_part=(uint8_t)a;
-	uint8_t fractional_part=(int)((a-(double)integer_part)*10.0);
+	uint8_t fractional_part=(int)(((double)a-(double)integer_part)*10.0);
 	sprintf(ftoa_str,"%d.%d",integer_part,fractional_part);
 	return ftoa_str;
